# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_handler.entryClearResponse:
    properties:
      deleted:
        type: integer
    type: object
  internal_handler.entryListResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/internal_handler.entryResponse'
        type: array
      hasMore:
        type: boolean
    type: object
  internal_handler.entryResponse:
    properties:
      author:
        type: string
      content:
        type: string
      createdAt:
        type: string
      feedId:
        type: string
      id:
        type: string
      publishedAt:
        type: string
      read:
        type: boolean
      readableContent:
        type: string
      starred:
        type: boolean
      thumbnailUrl:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
  internal_handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_handler.markAllReadRequest:
    properties:
      contentType:
        type: string
      feedId:
        type: string
      folderId:
        type: string
    type: object
...
  internal_handler.readableContentResponse:
    properties:
      readableContent:
        type: string
    type: object
...
  internal_handler.updateReadRequest:
    properties:
      read:
        type: boolean
    type: object
  internal_handler.updateStarredRequest:
    properties:
      starred:
        type: boolean
    type: object
...
info:
  contact: {}
  description: This is a modern RSS reader API.
  title: Gist API
  version: "1.0.1"
paths:
...
  /entries:
    get:
      description: Get a list of entries with optional filters and pagination
      parameters:
      - description: Filter by feed ID
        in: query
        name: feedId
        type: integer
      - description: Filter by folder ID
        in: query
        name: folderId
        type: integer
      - description: Filter by content type (article, picture, notification)
        in: query
        name: contentType
        type: string
      - description: Only return unread entries
        in: query
        name: unreadOnly
        type: boolean
      - description: Only return starred entries
        in: query
        name: starredOnly
        type: boolean
      - description: Limit the number of entries (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.entryListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: List entries
      tags:
      - entries
  /entries/{id}:
    get:
      description: Get a single entry by its ID
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.entryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Get entry
      tags:
      - entries
  /entries/{id}/fetch-readable:
    post:
      description: Extract readable content from the entry's original URL using readability
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.readableContentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Fetch readable content
      tags:
      - entries
  /entries/{id}/read:
    patch:
      consumes:
      - application/json
      description: Mark an entry as read or unread
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Read status
        in: body
        name: read
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateReadRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Update read status
      tags:
      - entries
  /entries/{id}/starred:
    patch:
      consumes:
      - application/json
      description: Mark an entry as starred or unstarred
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Starred status
        in: body
        name: starred
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateStarredRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Update starred status
      tags:
      - entries
  /entries/cache:
    delete:
      description: Delete all unstarred entries (preserves starred entries). Also
        resets all feeds' ETag/Last-Modified to force full refresh on next update.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.entryClearResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Clear entry cache
      tags:
      - entries
  /entries/mark-read:
    post:
      consumes:
      - application/json
      description: Mark all entries as read, optionally filtered by feed, folder,
        or content type
      parameters:
      - description: Filter criteria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.markAllReadRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Mark all as read
      tags:
      - entries
  /entries/readability-cache:
    delete:
      description: Delete all extracted readable content from entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.entryClearResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Clear readability cache
      tags:
      - entries
...
swagger: "2.0"